updates on TransactionController.php, migrations and API route

+++ mb40mktg/database/migrations/2018_08_29_124734_add_transaction_table.php
@@ -20,7 +20,7 @@ class AddTransactionTable extends Migration
                 $table->integer('account_id');
                 $table->integer('loan_id');
                 $table->integer('collector_id');
-                $table->integer('payment');
+                $table->float('payment');
                 $table->timestamps();
             });
         }
